auth-service/requirements.txt
# auth-service/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
python-jose
psycopg2-binary
python-dotenv
pytest